server/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
passlib
python-jose
typing-extensions
